# [kernel] Speed up TTY output processing