--- a/elks/fs/read_write.c
+++ b/elks/fs/read_write.c
@@ -90,7 +90,7 @@ int sys_read(unsigned int fd, char *buf, size_t count)
 	fop = file->f_op;
 	if (fop->read) {
 	    retval = (int) fop->read(file->f_inode, file, buf, count);
-	    schedule();
+	    schedule();         // FIXME removing these slows down localhost networking
 	}
     }
     return retval;
@@ -129,7 +129,7 @@ int sys_write(unsigned int fd, char *buf, size_t count)
 
 	    }
 	    written = (int) fop->write(inode, file, buf, count);
-	    schedule();         // FIXME should this be here?
+	    schedule();         // FIXME removing these slows down localhost networking
 	}
     }
     return written;