--- a/elks/include/linuxmt/chqueue.h
+++ b/elks/include/linuxmt/chqueue.h
@@ -1,7 +1,7 @@
-/* chqueue.h (C) 1997 Chad Page, rewritten Greg Haerr Oct 2020 */
-
 #ifndef __LINUXMT_CHQ_H
 #define __LINUXMT_CHQ_H
+
+/* chqueue.h (C) 1997 Chad Page, rewritten Greg Haerr Oct 2020 */
 
 struct ch_queue {
     unsigned char	*base;
@@ -11,15 +11,12 @@ struct ch_queue {
 };
 
 extern void chq_init(register struct ch_queue *,unsigned char *,int);
-/*extern void chq_erase(register struct ch_queue *);*/
 extern int chq_wait_wr(register struct ch_queue *,int);
+extern int chq_wait_rd(register struct ch_queue *,int);
 extern void chq_addch(register struct ch_queue *,unsigned char);
 extern void chq_addch_nowakeup(register struct ch_queue *,unsigned char);
-extern int chq_delch(register struct ch_queue *);
 extern int chq_peekch(register struct ch_queue *);
-/*extern int chq_full(register struct ch_queue *);*/
-
-extern int chq_wait_rd(register struct ch_queue *,int);
 extern int chq_getch(register struct ch_queue *);
+/*extern int chq_full(register struct ch_queue *);*/
 
 #endif