Tested on QEMU, this increases `sl` output by 25-50% on already fast systems. Discussed in https://github.com/ghaerr/elks/issues/1619#issuecomment-1712385485.

It seems @Vutshi has a knack for finding deep rabbit holes. Using the operation of `sl`, testing has found that the output speed varies considerably, based on something I can't pinpoint. I've looked at the scheduler, kernel timers, serial console, and all seem operating correctly. What is being observed is that (especially with this enhancement) `sl` runs fast ~0.45 seconds on QEMU on boot (from ~0.73s - 1.0s before). But after running networking, then stopping it, or running /bin/getty on a serial line, then stopping it, or even just typing ^D a few times to login: on serial, the `sl` output time increases 50% to 0.73s or more, then sometimes drops back to speedy again.

This needs testing on real hardware - I'm very interested to learn how much quicker this change gets us, as well as any slowdowns seen when `init 3` is run opening a /bin/getty on /dev/ttyS0.