@Vutshi 2023-09-11
@ghaerr I will test it as soon as I have access to the hardware. Meanwhile, I’ve learned about the new 8086 emulator which claims to be CPU cycle exact. 
https://int10h.org/blog/2023/07/martypc-pc-xt-emulator-raising-the-bar/
I wonder whether it can be useful in our context of performance debugging as it also provides some extensive debugging info.