@ghaerr 2023-09-11
>  I’ve learned about a new 8086 emulator which claims to be CPU cycle exact.

I just read about that emulator last week. IIRC it also emulates much of the PC hardware. I think it as well as other non-QEMU emulators would be very useful for ELKS. The problem with many emulators is they can take some time to setup properly. I have more emulators on my long list of things to do, but don't have lots of time to learn new ones. Anyone that has knowledge of a specific emulator or some time to set one up - it would be great to hear from you :)

@tyama501: I have `dosbox` setup and running which I use for testing your PC-98 port, but I have not been able to get it to boot non-FAT floppies (since we're using the its `boot` command running under DOS). Have you found a way to get it to boot an IBM PC MINIX-formattted floppy?

Are there other emulators that we should prioritize? What about Bochs, does anyone use that anymore? I have ongoing concerns about QEMU, as it has a nasty bug/feature that turns off the INT 8 hardware timer interrupt at times (I think too save CPU usage) which wreaks havoc on system timings and ELKS operation in general. It apparently only does this when the serial port is enabled. I would much like to test the `sl` timing issues on another emulator. Although the speed issues described above are likely ELKS internal, I haven't ruled out QEMU timing issues either.