@tyama501 2023-09-11
I didn't know DOSBox-X cannot boot non-FAT but I tried now and failed to boot...